requirements.txt
aiohttp>=1.0.0,<1.1.0
websockets>=3.1,<4.0
discord.py>=0.16.7,<1.0.0
